requirements.txt
# Core ML
numpy>=1.24
torch>=2.2
torchvision>=0.17
tensorflow>=2.15

# Computer vision
Pillow>=10.0
opencv-python>=4.8

# API server
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
pydantic>=2.4

# HTTP client (knowledge retrieval)
requests>=2.31

# Data ingestion
kagglehub>=0.1

# Config
pyyaml>=6.0

# Data / notebook / exploration
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
notebook>=7.0
tqdm>=4.66

# Testing
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.25

# Dev tooling
black>=23.0
isort>=5.12
